refactor: extract tasks page and components into modules
