+++ client/src/modules/Tasks/components/index.ts
@@ -1,0 +1,2 @@
+export * from './TableView/TableView';
+export * from './TaskModal/TaskModal';
