+++ client/src/modules/Tasks/pages/index.ts
@@ -1,0 +1,1 @@
+export * from './TasksPage/TasksPage';
