+++ client/src/modules/Tasks/types.ts
@@ -1,0 +1,5 @@
+import { TaskDto } from 'api';
+
+type ModalData = (Partial<Omit<TaskDto, 'attributes'>> & { attributes?: string }) | null;
+
+export type { ModalData };
